third_party/blink/web_tests/platform/generic/fast/dom/HTMLDocument/active-element-frames-expected.txt

When the contents of an iframe have focus, the activeElement in the parent document should be the iframe element.

  
Focusing top/input-0
PASS: top document.activeElement is top/input-0
PASS: top/iframe-1 document.activeElement is top/iframe-1/body-1
PASS: top/iframe-2 document.activeElement is top/iframe-2/body-2
PASS: top/iframe-1/iframe-3 document.activeElement is top/iframe-1/iframe-3/body-3
PASS: top/iframe-2/iframe-4 document.activeElement is top/iframe-2/iframe-4/body-4

Focusing top/iframe-1/input-1
PASS: top document.activeElement is top/iframe-1
PASS: top/iframe-1 document.activeElement is top/iframe-1/input-1
PASS: top/iframe-2 document.activeElement is top/iframe-2/body-2
PASS: top/iframe-1/iframe-3 document.activeElement is top/iframe-1/iframe-3/body-3
PASS: top/iframe-2/iframe-4 document.activeElement is top/iframe-2/iframe-4/body-4

Focusing top/iframe-2/input-2
PASS: top document.activeElement is top/iframe-2
PASS: top/iframe-1 document.activeElement is top/iframe-1/body-1
PASS: top/iframe-2 document.activeElement is top/iframe-2/input-2
PASS: top/iframe-1/iframe-3 document.activeElement is top/iframe-1/iframe-3/body-3
PASS: top/iframe-2/iframe-4 document.activeElement is top/iframe-2/iframe-4/body-4

Focusing top/iframe-1/iframe-3/input-3
PASS: top document.activeElement is top/iframe-1
PASS: top/iframe-1 document.activeElement is top/iframe-1/iframe-3
PASS: top/iframe-2 document.activeElement is top/iframe-2/body-2
PASS: top/iframe-1/iframe-3 document.activeElement is top/iframe-1/iframe-3/input-3
PASS: top/iframe-2/iframe-4 document.activeElement is top/iframe-2/iframe-4/body-4

Focusing top/iframe-2/iframe-4/input-4
PASS: top document.activeElement is top/iframe-2
PASS: top/iframe-1 document.activeElement is top/iframe-1/body-1
PASS: top/iframe-2 document.activeElement is top/iframe-2/iframe-4
PASS: top/iframe-1/iframe-3 document.activeElement is top/iframe-1/iframe-3/body-3
PASS: top/iframe-2/iframe-4 document.activeElement is top/iframe-2/iframe-4/input-4
